Rename JS parser build output file
