+++ .eslintignore
@@ -1,2 +1,2 @@
 **/node_modules/*
-parser/out.js
+parser/react-native-live-markdown-parser.js
